requirements.txt
compressai==1.2.6
huggingface-hub==0.16.4
matplotlib==3.5.1
natsort==8.4.0
numpy==1.22.3
opencv-python==4.6.0.66
ptflops==0.7
pytorch-msssim==1.0.0
sarpy==1.3.17
scipy==1.8.0
scikit-learn==1.0.2
scikit-image==0.19.2
thop==0.1.1.post2209072238
timm==0.9.5
torch-dct==0.1.6
torchmetrics==1.4.0
tqdm==4.64.0
